+++ .circleci/script/gen_ci.clj
@@ -151,7 +151,9 @@ java -jar \"$jar\" --config .build/bb.edn --deps-root . release-artifact \"$refl
                                  platform
                                  (if (= "aarch64" arch)
                                    "aarch64-"
-                                   ""))]
+                                   ""))
+        ;; The static musl binary cannot load shared libraries.
+        musl-static?     (and static? musl? (not= "aarch64" arch))]
     (gen-job shorted?
              (merge
               executor-conf
@@ -181,9 +183,9 @@ java -jar \"$jar\" --config .build/bb.edn --deps-root . release-artifact \"$refl
                                               (str base-install-cmd
                                                    "\nexport BABASHKA_TARBALL_CACHE=\"$HOME/.cache/babashka-tarballs\""
                                                    "\nmkdir -p \"$BABASHKA_TARBALL_CACHE\""
-                                                   (if (and static? musl? (not= "aarch64" arch))
+                                                   (if musl-static?
                                                      "\nsudo -E script/setup-musl"
-                                                     ;; non-musl linux builds link zlib statically
+                                                     ;; Other Linux builds link zlib statically.
                                                      "\nsudo -E script/setup-zlib"))))
                                           ;; after dev tools: the probe needs cc
                                           (when (not= "mac" platform)